+++ mods/rs2/src/draw.h
@@ -35,7 +35,7 @@ typedef struct {
 void add_draw_command(DRAW_COMMAND_TYPE type, void *data);
 void draw_hook(unsigned int unk);
 void draw_menu();
-void begin_warp(uint16_t level_index);
+void begin_warp(uint8_t level_index);
 void DrawLine(short x0, short y0, Color c0, short x1, short y1, Color c1);
 
 extern LevelData levels_table[NUM_LEVELS];
